2008-03-27  Felipe Erias Morandeira  <femorandeira@maemo.org>
        * main.h (struct package_info) : make installed_size field of type
        int64_t instead of int (N82350).


git-svn-id: https://garage.maemo.org/svn/hildon-app-mgr/trunk@1060 e55152ac-3a23-0410-980c-f88739b87167

+++ src/main.h
@@ -61,7 +61,7 @@ struct package_info {
   char *name;
   bool broken;
   char *installed_version;
-  int installed_size;
+  int64_t installed_size;
   char *installed_section;
   char *installed_pretty_name;
   char *available_version;
